Fix hover response
this wasn't quite matching the spec, and also wasn't consistent with the rest of the library
